:tada: v0.7.0 EpiRSV

+++ R/download.R
@@ -1,55 +1,39 @@
-
-
-
 #' Download from GISAID
 #'
 #' @param list_of_accession_ids list of accession_id from GISAID
 #' @return data.frame of complete data
-download <- function(credentials, list_of_accession_ids, get_sequence=FALSE) {
+download <- function(credentials, list_of_accession_ids, get_sequence=FALSE, clean_up=TRUE) {
   if (length(list_of_accession_ids) > 5000) {
     stop('Can only download a maxium of 5000 samples at a time.')
   }
   # select
-  accession_ids_string <- paste(list_of_accession_ids, collapse=", ")
+  message('Selecting entries...')
 
-  ev1 <- createCommand(
-    wid = credentials$wid,
-    pid = credentials$selection_pid,
-    cid = credentials$panel_cid,
-    cmd = 'setTarget',
-    params = list(cvalue=accession_ids_string, ceid=credentials$selection_ceid), #hack for empty {}
-    equiv = paste0("ST", credentials$selection_ceid)
-  )
+  response <- select_entries(credentials = credentials, list_of_accession_ids=list_of_accession_ids)
 
-  ev2 <- createCommand(
-    wid = credentials$wid,
-    pid = credentials$selection_pid,
-    cid = credentials$panel_cid,
-    cmd = 'ChangeValue',
-    params = list(cvalue=accession_ids_string, ceid=credentials$selection_ceid), #hack for empty {}
-    equiv = paste0("CV", credentials$selection_ceid)
-  )
-  ev3 <- createCommand(
-    wid = credentials$wid,
-    pid = credentials$selection_pid,
-    cid = credentials$panel_cid,
-    cmd = 'OK',
-    params = setNames(list(), character(0)) #hack for empty {}
-  )
-  json_queue <- list(queue = list(ev1, ev2, ev3))
-  data <- formatDataForRequest(credentials$sid, credentials$wid, credentials$selection_pid, json_queue, timestamp())
-  res <-
-    httr::POST(GISAID_URL, httr::add_headers(.headers = headers), body = data)
+  download_cmd <- 'ToolDownload'
+  download_pid_wid <- list(pid=credentials$pid, wid=credentials$wid)
+  if (credentials$database == 'EpiRSV') {
+    download_cmd <- 'Download'
+    download_pid_wid <- get_download_panel(credentials$sid, credentials$wid, credentials$pid, credentials$query_cid)
+    #load panel
+    downloal_page <-
+      send_request(paste0('sid=', credentials$sid, '&pid=', download_pid_wid$pid))
+
+    download_page_text = httr::content(downloal_page, as = 'text')
+    credentials$download_panel_cid <- extract_first_match("'(.{5,20})','RSVDownloadSelectionComponent", download_page_text)
+    #send_back_cmd(credentials$sid, download_pid_wid$wid, download_pid_wid$pid, credentials$download_panel_cid)
+  }
 
   ev <- createCommand(
-    wid = credentials$wid,
-    pid = credentials$pid,
-    cid = credentials$download_cid,
-    cmd = 'ToolDownload',
+    wid = download_pid_wid$wid,
+    pid = download_pid_wid$pid,
+    cid = credentials$download_panel_cid,
+    cmd = download_cmd,
     params = setNames(list(), character(0)) #hack for empty {}
   )
   json_queue <- list(queue = list(ev))
-  data <- formatDataForRequest(credentials$sid, credentials$wid, credentials$pid, json_queue, timestamp())
+  data <- formatDataForRequest(credentials$sid, download_pid_wid$wid, download_pid_wid$pid, json_queue, timestamp())
   message('Compressing data. Please wait...')
   res <-
     httr::POST(GISAID_URL, httr::add_headers(.headers = headers), body = data)
@@ -67,13 +51,20 @@ download <- function(credentials, list_of_accession_ids, get_sequence=FALSE) {
       Sys.sleep(1)
     }
   }
+  log.debug(j)
   # get download link
   message('Data ready.')
-  ev$cmd <- "generateDownloadDone"
+  ev <- createCommand(
+    wid = credentials$wid,
+    pid = credentials$pid,
+    cid = credentials$query_cid,
+    cmd = "generateDownloadDone",
+    params = setNames(list(), character(0)) #hack for empty {}
+  )
   json_queue <- list(queue = list(ev))
   data <- formatDataForRequest(credentials$sid, credentials$wid, credentials$pid, json_queue, timestamp())
   res <-
-    httr::POST(GISAID_URL, httr::add_headers(.headers = headers), body = data)
+    send_request(method = 'POST', data=data)
   j <- parseResponse(res)
 
   # extract download url
@@ -81,35 +72,50 @@ download <- function(credentials, list_of_accession_ids, get_sequence=FALSE) {
   # download zip
   tryCatch({
     message('Downloading...')
-    download.file(download_url, "gisaidr_data_tmp.tar", quiet = TRUE, method = 'auto', )
-    # unzip
-    untar("gisaidr_data_tmp.tar", exdir="gisaidr_data_tmp", restore_times = FALSE)
-    # load into df
-    metadataFile <- list.files("gisaidr_data_tmp", pattern = "*.metadata.tsv")[1]
-    if (is.na(metadataFile)) {
-      message("gisaid_data files:")
-      message(list.files("gisaidr_data_tmp"))
-      stop("Could not find metadata file.")
-    }
-    df <- read.csv(paste0("gisaidr_data_tmp/", metadataFile), sep="\t", quote="")
-    df <- df[order(df$gisaid_epi_isl, decreasing = TRUE),]
-    colnames(df)[3] <- "accession_id"
-    if (get_sequence) {
-      # join sequence
-      sequencesFile <- list.files("gisaidr_data_tmp", pattern = "*.sequences.fasta")[1]
-      if (is.na(sequencesFile)) {
-        stop("Could not find sequences file.")
+
+    if (credentials$database == "EpiCoV") {
+      download.file(download_url, "gisaidr_data_tmp.tar", quiet = TRUE, method = 'auto', )
+      # unzip
+      untar("gisaidr_data_tmp.tar", exdir="gisaidr_data_tmp", restore_times = FALSE)
+      # load into df
+      metadataFile <- list.files("gisaidr_data_tmp", pattern = "*.metadata.tsv")[1]
+      if (is.na(metadataFile)) {
+        message("gisaid_data files:")
+        message(list.files("gisaidr_data_tmp"))
+        stop("Could not find metadata file.")
       }
-      seq_df <- read_fasta(paste0("gisaidr_data_tmp/", sequencesFile))
-      df <- merge(x = df, y = seq_df, by = "strain", all = TRUE)
+      df <- read.csv(paste0("gisaidr_data_tmp/", metadataFile), sep="\t", quote="")
+      df <- df[order(df$gisaid_epi_isl, decreasing = TRUE),]
+      colnames(df)[3] <- "accession_id"
+      if (get_sequence) {
+        # join sequence
+        sequencesFile <- list.files("gisaidr_data_tmp", pattern = "*.sequences.fasta")[1]
+        if (is.na(sequencesFile)) {
+          stop("Could not find sequences file.")
+        }
+        seq_df <- read_fasta(paste0("gisaidr_data_tmp/", sequencesFile))
+        df <- merge(x = df, y = seq_df, by = "strain", all = TRUE)
+      }
+    } else {
+      sequencesFile <- "gisaidr_data_tmp.fasta"
+      download.file(download_url, sequencesFile, quiet = TRUE, method = 'auto')
+      fdf <- read_fasta(sequencesFile, get_sequence=get_sequence)
+      df <- merge(data.frame(do.call('rbind', strsplit(as.character(fdf$strain), '|', fixed=TRUE))), fdf)
+      if (get_sequence) {
+        names(df) <- c('strain', 'accession_id', ' collection_date', 'description', 'sequence')
+      }
+      names(df) <- c('strain', 'accession_id', ' collection_date', 'description')
     }
+
   }, finally = {
     # clean up
-    if (file.exists("gisaidr_data_tmp.tar")) {
-      file.remove("gisaidr_data_tmp.tar")
-    }
-    if (file.exists("gisaidr_data_tmp/")) {
-      unlink("gisaidr_data_tmp/", recursive = TRUE)
+    if (clean_up){
+      if (file.exists("gisaidr_data_tmp.tar")) {
+        file.remove("gisaidr_data_tmp.tar")
+      }
+      if (file.exists("gisaidr_data_tmp/")) {
+        unlink("gisaidr_data_tmp/", recursive = TRUE)
+      }
     }
   })
   # replace ? with NA
